strawberry_graphql/requirements.txt
Django~=3.2.5

django-cors-headers
strawberry-graphql-django
